Do not allow insolvency resource to be updated if corresponding transaction has already been closed

+++ handlers/attachment_resource.go
@@ -27,6 +27,21 @@ func HandleSubmitAttachment(svc dao.Service) http.Handler {
 
 		log.InfoR(req, fmt.Sprintf("start POST request for submit attachment with transaction id: %s", transactionID))
 
+		// Check if transaction is closed
+		isTransactionClosed, err, httpStatus := service.CheckIfTransactionClosed(transactionID, req)
+		if err != nil {
+			log.ErrorR(req, fmt.Errorf("error checking transaction status for [%v]: [%s]", transactionID, err))
+			m := models.NewMessageResponse(fmt.Sprintf("error checking transaction status for [%v]: [%s]", transactionID, err))
+			utils.WriteJSONWithStatus(w, req, m, httpStatus)
+			return
+		}
+		if isTransactionClosed {
+			log.ErrorR(req, fmt.Errorf("transaction [%v] is already closed and cannot be updated", transactionID))
+			m := models.NewMessageResponse(fmt.Sprintf("transaction [%v] is already closed and cannot be updated", transactionID))
+			utils.WriteJSONWithStatus(w, req, m, httpStatus)
+			return
+		}
+
 		attachmentType := req.FormValue("attachment_type")
 
 		file, header, err := req.FormFile("file")
@@ -262,6 +277,21 @@ func HandleDeleteAttachment(svc dao.Service) http.Handler {
 		}
 
 		log.InfoR(req, fmt.Sprintf("start DELETE request for attachment with transaction id: %s, attachment id: %s", transactionID, attachmentID))
+
+		// Check if transaction is closed
+		isTransactionClosed, err, httpStatus := service.CheckIfTransactionClosed(transactionID, req)
+		if err != nil {
+			log.ErrorR(req, fmt.Errorf("error checking transaction status for [%v]: [%s]", transactionID, err))
+			m := models.NewMessageResponse(fmt.Sprintf("error checking transaction status for [%v]: [%s]", transactionID, err))
+			utils.WriteJSONWithStatus(w, req, m, httpStatus)
+			return
+		}
+		if isTransactionClosed {
+			log.ErrorR(req, fmt.Errorf("transaction [%v] is already closed and cannot be updated", transactionID))
+			m := models.NewMessageResponse(fmt.Sprintf("transaction [%v] is already closed and cannot be updated", transactionID))
+			utils.WriteJSONWithStatus(w, req, m, httpStatus)
+			return
+		}
 
 		responseType, err := service.DeleteAttachment(attachmentID, req)
 		if err != nil {
